v0.06: Update

+++ src/MySQLCore.API/Controllers/CRUDTransactionController.cs
@@ -1,16 +1,16 @@
 using Microsoft.AspNetCore.Mvc;
 using MySQLCore.Core.CoreHelpers;
-using MySQLCore.Core.DTOs;
 using MySQLCore.Core.Interfaces.InterfaceServices;
+using MySQLCore.Core.Models.DTOs;
 
 namespace MySQLCore.API.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class CRUDTransactionController : ControllerBase
+    public class CRUDTransactionController : BaseController
     {
         private readonly ICRUDTransactionService _service;
-        public CRUDTransactionController(ICRUDTransactionService service)
+        public CRUDTransactionController(ICRUDTransactionService service, ILogger<CRUDTransactionController> logger) : base(logger)
         {
             _service = service ?? throw new ArgumentNullException(nameof(service));
         }
@@ -25,13 +25,18 @@ namespace MySQLCore.API.Controllers
         [HttpGet]
         [Route("GetRecordById/{Id:int}")]
         public async Task<ActionResult<CRUDTransactionDTO>> GetRecordById(int Id){
-            var result = await _service.GetRecordById(Id);
-            return result.NullChecker() ? result : new CRUDTransactionDTO();
+            if (Id > 0)  {
+                var result = await _service.GetRecordById(Id);
+                return result.NullChecker() ? result : new CRUDTransactionDTO();
+            }
+            
+            return BadRequest(); 
         }
 
         [HttpPost]
         [Route("CreateRecord")]
-        public async Task<ActionResult<bool>> CreateRecord(CRUDTransactionDTO dTO){
+        public async Task<ActionResult<bool>> CreateRecord(CRUDTransactionDTO dTO) {
+            if (!ModelState.IsValid) { return BadRequest(); }
             var result = await _service.CreateRecord(dTO);
             return result;
         }
@@ -39,6 +44,7 @@ namespace MySQLCore.API.Controllers
         [HttpPut]
         [Route("UpdateRecord")]
         public async Task<ActionResult<bool>> UpdateRecord(CRUDTransactionDTO dTO){
+            if (!ModelState.IsValid) { return BadRequest(); }
             var result = await _service.UpdateRecord(dTO);
             return result;
         }
@@ -46,8 +52,12 @@ namespace MySQLCore.API.Controllers
         [HttpDelete]
         [Route("DeleteRecord")]
         public async Task<ActionResult<bool>> DeleteRecord(int id){
-            var result = await _service.DeleteRecord(id);
-            return result;
+            if (id > 0) {
+                var result = await _service.DeleteRecord(id);
+                return result;
+            }
+            
+            return BadRequest();
         }
     }
 }
